# Early out in ObjectPool to avoid loop execution