This is a micro optimization which caught my eye when profiling my own application. I haven't tested the impact on overall execution performance. But given that the pool is used at basically every function invocation, I assume that it won't hurt to change it. Given that the change doesn't make the code much harder to understand or maintain, I just did it.

---

## Motivation

The object pool is only used internally and a few of them are allocated per `Engine` instance. The pool has basically "best effort" guarantees. Given that `Engine` itself is not declared as thread safe, no synchronization was used for the change.

The performance problem was mainly with empty or almost empty pools. In those cases, the pool always looped over its internal array to find a possible match. Given that the pool can easily track how many items are currently stored within the array, an early out mechanism was added.

## Benchmark

Here are the results. The benchmark code is attached below.

GetFromEmpty:
| Method               | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
| Old                  | 82.46 ns | 1.503 ns | 1.333 ns | 0.0253 |     424 B |
| New                  | 49.75 ns | 0.972 ns | 1.157 ns | 0.0258 |     432 B |

GetFromFilledWithOne:
| Method               | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
| Old                  | 95.70 ns | 1.226 ns | 1.147 ns | 0.0253 |     424 B |
| New                  | 53.81 ns | 1.053 ns | 1.952 ns | 0.0258 |     432 B |

GetFromAlmostFull:
| Method               | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
| Old                  | 83.84 ns | 1.189 ns | 1.112 ns | 0.0253 |     424 B |
| New                  | 82.10 ns | 0.813 ns | 0.760 ns | 0.0257 |     432 B |

GetFromFullyFilled:
| Method               | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
| Old                  | 97.51 ns | 1.073 ns | 0.951 ns | 0.0267 |     448 B |
| New                  | 91.66 ns | 0.930 ns | 0.870 ns | 0.0272 |     456 B |

<details>
  <summary>Benchmark Code</summary>
  
  The benchmark always fetches 12 items from the pool. The overhead of allocating and
  filling the pool is not relevant because the same code is used for both, before and after.

  ```cs
  [MemoryDiagnoser]
  [GroupBenchmarksBy(BenchmarkLogicalGroupRule.ByCategory, BenchmarkLogicalGroupRule.ByParams)]
  public class ObjectPoolBenchmark
  {
      [Benchmark()]
      public void GetFromEmpty()
      {
          var pool = new ObjectPool<object>(Allocator, size: 10);
          AllocateItems(pool);
      }
  
      [Benchmark()]
      public void GetFromFilledWithOne()
      {
          var pool = new ObjectPool<object>(Allocator, size: 10);
          pool.Free(Allocator());
          AllocateItems(pool);
      }
  
      [Benchmark()]
      public void GetFromAlmostFull()
      {
          var pool = new ObjectPool<object>(Allocator, size: 10);
          for (int i = 0; i < 9; i++)
              pool.Free(Allocator());
  
          AllocateItems(pool);
      }
  
      [Benchmark()]
      public void GetFromFullyFilled()
      {
          var pool = new ObjectPool<object>(Allocator, size: 10);
          for (int i = 0; i < 11; i++)
              pool.Free(Allocator());
  
          AllocateItems(pool);
      }
  
      private static void AllocateItems(ObjectPool<object> pool)
      {
          for (int i = 0; i < 12; i++)
              pool.Allocate();
      }
  
      private static object Allocator()
          => new object();
  }
  ```
</details>