@lahma 2024-10-24
Great find and extensive benchmarking! The current implementation was based on old object pool code and since that it has been improved, I wonder if changes in https://github.com/dotnet/aspnetcore/pull/45251 would reflect well on Jint side too...