@tomatosalat0 2024-10-24
Good idea! I tried to use it but the results do not look great.

Microsoft.Extensions.ObjectPool:
| Method               | Mean     | Error   | StdDev  | Gen0   | Gen1   | Allocated |
|--------------------- |---------:|--------:|--------:|-------:|-------:|----------:|
| GetFromEmpty         | 171.0 ns | 3.08 ns | 2.88 ns | 0.0811 | 0.0002 |   1.33 KB |
| GetFromFilledWithOne | 171.5 ns | 2.58 ns | 2.29 ns | 0.0811 | 0.0002 |   1.33 KB |
| GetFromAlmostFull    | 220.2 ns | 3.86 ns | 3.61 ns | 0.0811 | 0.0002 |   1.33 KB |
| GetFromFullyFilled   | 223.5 ns | 1.98 ns | 1.66 ns | 0.0811 | 0.0002 |   1.33 KB |

Which is not a big surprise because the this pool does support multi threading, which ours doesn't.
However, I took a look into their implementation and tried to replicate it - without concurrency support:

* Usage of `System.Collections.Generic.Queue<T>` instead of the array with a pre-allocated capacity
* Ensuring that the queue never gets more than N items.
* Keep the usage of `_firstItem` as it is what they have done as well

Well, the results were better, but not as good as the PR version.

Queue&lt;T&gt; and _firstItem:
| Method               | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
| GetFromEmpty         | 60.51 ns | 1.228 ns | 1.088 ns | 0.0291 |     488 B |
| GetFromFilledWithOne | 63.87 ns | 1.285 ns | 2.537 ns | 0.0291 |     488 B |
| GetFromAlmostFull    | 97.33 ns | 0.769 ns | 0.719 ns | 0.0291 |     488 B |
| GetFromFullyFilled   | 95.53 ns | 0.806 ns | 0.715 ns | 0.0291 |     488 B |

In summary: I think the stupid simple version of the Pool within the code is actually not that bad.