@sebastienros 2024-10-24
Pretty sure I have this pool implementation in other projects ... since it's taken from the dotnet runtime there might be a PR to do there also.